requirements.txt
crewai==0.22.5
crewai-tools==0.0.15
python-dotenv==1.0.0
langchain-community===0.0.31